requirements.txt
# Core PDF Processing
pymupdf==1.23.25  # Lightweight PDF text/tables
pdf2image==1.17.0  # Scanned PDFs
#poppler-utils==0.68.0  # Required for pdf2image (will use chocolatey install)

# Layout Analysis (Donut replaces Detectron2)
layoutparser==0.3.4  # With Donut model
transformers==4.41.2  # For Donut

# OCR
pytesseract==0.3.10
easyocr==1.7.1  # Fallback
tesseract-ocr  # Windows executable (via chocolatey)

# Image Processing
opencv-python-headless==4.9.0.80  # Lighter than full OpenCV
Pillow==10.3.0  # Image handling
pdf2image==1.17.0  # For scanned PDFs

# Utilities
numpy==1.26.4  # Latest stable before v2
pandas==2.2.2  # Data structuring
tqdm==4.66.2  # Progress bars
python-dotenv==1.0.1  # Path management
PyYAML==6.0.2  # YAML parsing

# Error Handling
tenacity==8.2.3  # Retries
loguru==0.7.2  # Simplified logging